requirements.txt
certifi==2020.11.8
click==7.1.2
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
plotly @ file:///tmp/build/80754af9/plotly_1605121809988/work
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
six @ file:///C:/ci/six_1605205404068/work
Werkzeug==1.0.1
wincertstore==0.2
